[#394] Document SampleMutUninit
