+++ iceoryx2/src/sample_mut.rs
@@ -25,9 +25,8 @@
 //! #
 //! # let publisher = service.publisher_builder().create()?;
 //!
-//! let sample = publisher.loan_uninit()?;
-//! // write 1234 into sample
-//! let mut sample = sample.write_payload(1234);
+//! // initializes the payload with `Default::default()`
+//! let sample = publisher.loan()?;
 //! // override contents with 456 because its fun
 //! *sample.payload_mut() = 456;
 //!
@@ -52,9 +51,8 @@
 //! # let publisher = service.publisher_builder().max_slice_len(16).create()?;
 //!
 //! let slice_length = 12;
-//! let sample = publisher.loan_slice_uninit(slice_length)?;
-//! // initialize the n-th element of the slice with n * 1234
-//! let mut sample = sample.write_from_fn(|n| n * 1234);
+//! // initializes every element of the slice with `Default::default()`
+//! let sample = publisher.loan_slice(slice_length)?;
 //! // override the content of the first element with 42
 //! sample.payload_mut()[0] = 42;
 //!
@@ -73,15 +71,12 @@ use crate::{
 use iceoryx2_cal::shared_memory::*;
 use std::{
     fmt::{Debug, Formatter},
-    mem::MaybeUninit,
     sync::Arc,
 };
 
 /// Acquired by a [`crate::port::publisher::Publisher`] via
 ///  * [`crate::port::publisher::Publisher::loan()`],
-///  * [`crate::port::publisher::Publisher::loan_uninit()`]
 ///  * [`crate::port::publisher::Publisher::loan_slice()`]
-///  * [`crate::port::publisher::Publisher::loan_slice_uninit()`]
 ///
 /// It stores the payload that will be sent
 /// to all connected [`crate::port::subscriber::Subscriber`]s. If the [`SampleMut`] is not sent
@@ -91,12 +86,9 @@ use std::{
 ///
 /// Does not implement [`Send`] since it releases unsent samples in the [`crate::port::publisher::Publisher`] and the
 /// [`crate::port::publisher::Publisher`] is not thread-safe!
-///
-/// The generic parameter `M` is either a `Payload` or a [`core::mem::MaybeUninit<Payload>`], depending
-/// which API is used to obtain the sample.
 pub struct SampleMut<Service: crate::service::Service, Payload: Debug + ?Sized, UserHeader> {
-    data_segment: Arc<DataSegment<Service>>,
-    ptr: RawSampleMut<Header, UserHeader, Payload>,
+    pub(crate) data_segment: Arc<DataSegment<Service>>,
+    pub(crate) ptr: RawSampleMut<Header, UserHeader, Payload>,
     pub(crate) offset_to_chunk: PointerOffset,
 }
 
@@ -121,186 +113,6 @@ impl<Service: crate::service::Service, Payload: Debug + ?Sized, UserHeader> Drop
 {
     fn drop(&mut self) {
         self.data_segment.return_loaned_sample(self.offset_to_chunk);
-    }
-}
-
-impl<Service: crate::service::Service, Payload: Debug, UserHeader>
-    SampleMut<Service, MaybeUninit<Payload>, UserHeader>
-{
-    pub(crate) fn new(
-        data_segment: &Arc<DataSegment<Service>>,
-        ptr: RawSampleMut<Header, UserHeader, MaybeUninit<Payload>>,
-        offset_to_chunk: PointerOffset,
-    ) -> Self {
-        Self {
-            data_segment: Arc::clone(data_segment),
-            ptr,
-            offset_to_chunk,
-        }
-    }
-}
-
-impl<Service: crate::service::Service, Payload: Debug, UserHeader>
-    SampleMut<Service, [MaybeUninit<Payload>], UserHeader>
-{
-    pub(crate) fn new(
-        data_segment: &Arc<DataSegment<Service>>,
-        ptr: RawSampleMut<Header, UserHeader, [MaybeUninit<Payload>]>,
-        offset_to_chunk: PointerOffset,
-    ) -> Self {
-        Self {
-            data_segment: Arc::clone(data_segment),
-            ptr,
-            offset_to_chunk,
-        }
-    }
-}
-
-impl<Service: crate::service::Service, Payload: Debug, UserHeader>
-    SampleMut<Service, MaybeUninit<Payload>, UserHeader>
-{
-    /// Writes the payload to the sample and labels the sample as initialized
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let node = NodeBuilder::new().create::<ipc::Service>()?;
-    /// #
-    /// # let service = node.service_builder(&"My/Funk/ServiceName".try_into()?)
-    /// #     .publish_subscribe::<u64>()
-    /// #     .open_or_create()?;
-    /// #
-    /// # let publisher = service.publisher_builder().create()?;
-    ///
-    /// let sample = publisher.loan_uninit()?;
-    /// let sample = sample.write_payload(1234);
-    ///
-    /// sample.send()?;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn write_payload(mut self, value: Payload) -> SampleMut<Service, Payload, UserHeader> {
-        self.payload_mut().write(value);
-        // SAFETY: this is safe since the payload was initialized on the line above
-        unsafe { self.assume_init() }
-    }
-
-    /// Extracts the value of the [`core::mem::MaybeUninit<Payload>`] container and labels the sample as initialized
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that [`core::mem::MaybeUninit<Payload>`] really is initialized. Calling this when
-    /// the content is not fully initialized causes immediate undefined behavior.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let node = NodeBuilder::new().create::<ipc::Service>()?;
-    /// #
-    /// # let service = node.service_builder(&"My/Funk/ServiceName".try_into()?)
-    /// #     .publish_subscribe::<u64>()
-    /// #     .open_or_create()?;
-    /// #
-    /// # let publisher = service.publisher_builder().create()?;
-    ///
-    /// let mut sample = publisher.loan_uninit()?;
-    /// sample.payload_mut().write(1234);
-    /// let sample = unsafe { sample.assume_init() };
-    ///
-    /// sample.send()?;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub unsafe fn assume_init(self) -> SampleMut<Service, Payload, UserHeader> {
-        // the transmute is not nice but safe since MaybeUninit is #[repr(transparent)] to the inner type
-        std::mem::transmute(self)
-    }
-}
-
-impl<Service: crate::service::Service, Payload: Debug, UserHeader>
-    SampleMut<Service, [MaybeUninit<Payload>], UserHeader>
-{
-    /// Extracts the value of the slice of [`core::mem::MaybeUninit<Payload>`] and labels the sample as initialized
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that every element of the slice of [`core::mem::MaybeUninit<Payload>`]
-    /// is initialized. Calling this when the content is not fully initialized causes immediate undefined behavior.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// use core::mem::MaybeUninit;
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let node = NodeBuilder::new().create::<ipc::Service>()?;
-    /// #
-    /// # let service = node.service_builder(&"My/Funk/ServiceName".try_into()?)
-    /// #     .publish_subscribe::<[usize]>()
-    /// #     .open_or_create()?;
-    /// #
-    /// # let publisher = service.publisher_builder().max_slice_len(32).create()?;
-    ///
-    /// let slice_length = 10;
-    /// let mut sample = publisher.loan_slice_uninit(slice_length)?;
-    ///
-    /// for element in sample.payload_mut() {
-    ///     element.write(1234);
-    /// }
-    ///
-    /// let sample = unsafe { sample.assume_init() };
-    ///
-    /// sample.send()?;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub unsafe fn assume_init(self) -> SampleMut<Service, [Payload], UserHeader> {
-        // the transmute is not nice but safe since MaybeUninit is #[repr(transparent)] to the inner type
-        std::mem::transmute(self)
-    }
-
-    /// Writes the payload to the sample and labels the sample as initialized
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// # let node = NodeBuilder::new().create::<ipc::Service>()?;
-    /// #
-    /// # let service = node.service_builder(&"My/Funk/ServiceName".try_into()?)
-    /// #     .publish_subscribe::<[usize]>()
-    /// #     .open_or_create()?;
-    /// #
-    /// # let publisher = service.publisher_builder().max_slice_len(16).create()?;
-    ///
-    /// let slice_length = 12;
-    /// let sample = publisher.loan_slice_uninit(slice_length)?;
-    /// let sample = sample.write_from_fn(|n| n + 123);
-    ///
-    /// sample.send()?;
-    ///
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn write_from_fn<F: FnMut(usize) -> Payload>(
-        mut self,
-        mut initializer: F,
-    ) -> SampleMut<Service, [Payload], UserHeader> {
-        for (i, element) in self.payload_mut().iter_mut().enumerate() {
-            element.write(initializer(i));
-        }
-
-        // SAFETY: this is safe since the payload was initialized on the line above
-        unsafe { self.assume_init() }
     }
 }
